# oas/OAS.yaml
openapi: 3.1.0
info:
  title: Multi-Tenant Salon API
  description: |
    REST API for the multi-tenant SaaS salon/salon management platform.

    ## Module / Access Hierarchy

    The API is split into four distinct path namespaces:

    | Namespace | Base Path | Who uses it |
    |---|---|---|
    | **Salon Onboarding** | `/api/salon-onboarding` | Anyone registering a new salon tenant; platform-level listing |
    | **Customer** | `/api/salon/{salonId}/...` | Public/customer-facing — browse, book, read |
    | **Admin** | `/api/salon-admin/{salonId}/...` | Salon owner/administrator — manage, configure, operate. Also callable by a platform super-admin for any salon, since super-admin management of a specific salon reuses these endpoints rather than duplicating them. |
    | **Super Admin** | `/api/salon-super-admin/...` | Platform super-admin — cross-tenant management of all salons |
    | **Staff Portal** | `/api/salon-staff/...` | Authenticated staff member — self-service profile, appointments, personal holidays |
    | **Utility** | `/api/salon-utility/...` | Any consumer needing reference data (countries with embedded currency info) |

    ### Customer sub-paths
    - `/api/salon/{salonId}/services/...` — browse the service catalog
    - `/api/salon/{salonId}/staff/...` — view staff members
    - `/api/salon/{salonId}/booking/...` — slot discovery and appointment creation
    - `/api/salon/{salonId}/website` — public website theme
...

    ### Admin sub-paths
...

...
servers:
...
paths:
...
  /api/salon-onboarding:
    get:
      tags: [SalonOnboarding]
      summary: List all salons (platform view)
      operationId: listSalons
      responses:
        "200":
          description: List of all salons across all tenants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Salon"

    post:
      tags: [SalonOnboarding]
      summary: Register a new salon (tenant onboarding)
      operationId: createSalon
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSalonRequest"
      responses:
        "201":
          description: Salon registered; owner auto-enrolled as staff member
          headers:
            Location:
              description: URL of the newly created salon, e.g. /api/salon/{id}
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateSalonResponse"

  # ─── Customer — Salon Discovery ─────────────────────────────────────────────
  /api/salon/{salonIdOrHandler}:
    get:
      tags: [Salon]
      summary: Get salon by ID or handler (public)
      description: |
        Accepts either a UUID (`a1b2c3d4-e5f6-7890-abcd-ef1234567890`) or a handler slug
        (`glam-salon`). UUID lookup is tried first; if the value is not a valid UUID the
        request falls back to a handler lookup.
      operationId: getSalon
      parameters:
        - name: salonIdOrHandler
          in: path
          required: true
          schema:
            type: string
          description: Salon UUID or URL-friendly handler slug
      responses:
        "200":
          description: Salon details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Salon"
        "404":
          $ref: "#/components/responses/NotFound"

  # ─── Customer — Services ──────────────────────────────────────────────────────
...
  /api/salon-admin/my-salons:
    get:
      tags: [SalonAdmin]
      summary: Look up the caller's own salons (admin login)
      description: |
        Returns all salons owned by the authenticated caller.
        Used by the admin login flow to resolve the salonId before entering the admin UI.
        - **1 result** → proceed directly to the salon-admin UI.
        - **2+ results** → show a salon-picker screen and let the owner choose.
        - **0 results** → show "no salon found" error.

        The owner's identity is taken from the `sub` claim of the bearer token —
        never from client input, since that would let a caller ask for someone
        else's salons. The `email` query parameter is only a fallback for
        unauthenticated local/mock-mode use and is ignored whenever a bearer
        token is present.
      operationId: getMySalons
      security:
        - BearerAuth: []
        - {}
      parameters:
        - name: email
          in: query
          required: false
          schema:
            type: string
            format: email
          example: owner@example.com
          description: |
            Fallback owner email, used only when no bearer token is present
            (local/mock-mode login). Ignored when the caller is authenticated.
      responses:
        "200":
          description: One or more salons found for this email
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Salon"
        "401":
          description: No bearer token and no email fallback provided
        "404":
          $ref: "#/components/responses/NotFound"

  # ─── Admin — Salon Management ────────────────────────────────────────────────
  /api/salon-admin/{salonId}:
    parameters:
      - $ref: "#/components/parameters/SalonId"

    get:
      tags: [SalonAdmin]
      summary: Get salon details (admin)
      description: Accepts either a UUID or a handler slug — same lookup logic as the public endpoint.
      operationId: adminGetSalon
      responses:
        "200":
          description: Salon details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Salon"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags: [SalonAdmin]
      summary: Update salon details (admin)
      description: |
        Admin-only. Updates name, location, contact info, and operating hours. Handler, features, and owner are never changed here — use PUT /api/salon-super-admin/salons/{id}/owner to change the owner.
        Emits **SalonUpdatedEvent** → owner receives a notification email.
      operationId: updateSalon
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSalonRequest"
      responses:
        "200":
          description: Updated salon
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Salon"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags: [SalonAdmin]
      summary: Delete a salon (admin)
      description: Admin-only. Permanently removes the salon and all associated data (services, staff, bookings) via cascade delete.
      operationId: deleteSalon
      responses:
        "204":
          description: Salon deleted
        "404":
          $ref: "#/components/responses/NotFound"
  /api/salon-admin/{salonId}/features:
    parameters:
      - $ref: "#/components/parameters/SalonId"

    put:
      tags: [SalonAdmin]
      summary: Replace salon feature flags (admin)
      description: |
        Admin-only. Replaces the full set of enabled features for the salon (BOOKING, ANALYTICS, etc.).
        Emits **SalonUpdatedEvent** → owner receives a notification email.
      operationId: updateSalonFeatures
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/SalonFeature"
              example: ["BOOKING", "ANALYTICS"]
      responses:
        "200":
          description: Features updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Salon"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/salon-admin/{salonId}/booking-settings:
    parameters:
      - $ref: "#/components/parameters/SalonId"

    patch:
      tags: [SalonAdmin]
      summary: Update booking settings (admin)
      description: |
        Admin-only. Partially updates only the booking-related salon settings —
        `bookingAdvanceDays` and `bookingRequiresConfirmation`. All other salon fields
        are preserved. Either field may be omitted; omitted fields keep their current value.
        Emits **SalonUpdatedEvent** → owner receives a notification email.
      operationId: patchBookingSettings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchBookingSettingsRequest"
      responses:
        "200":
          description: Settings updated; full salon object returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Salon"
        "404":
          $ref: "#/components/responses/NotFound"

  # ─── Admin — Holidays ────────────────────────────────────────────────────────
  /api/salon-admin/{salonId}/holidays:
    parameters:
      - $ref: "#/components/parameters/SalonId"

    get:
      tags: [AdminHolidays]
      summary: List salon holidays (admin)
      operationId: listHolidays
      responses:
        "200":
          description: List of holidays
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SalonHoliday"

    post:
      tags: [AdminHolidays]
      summary: Add a holiday (admin)
      description: |
        Admin-only. Creates a holiday record and automatically generates salon-closure rows
        for the current year through currentYear+4 (for recurring holidays) or for the
        specified year only (for one-off holidays). Invalid dates (e.g. Feb 29 on a non-leap
        year) are skipped silently.
      operationId: addHoliday
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddHolidayRequest"
      responses:
        "200":
          description: Holiday created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalonHoliday"
  /api/salon-admin/{salonId}/holidays/{holidayId}:
    parameters:
      - $ref: "#/components/parameters/SalonId"
      - name: holidayId
        in: path
        required: true
        schema:
          type: integer
          format: int64
        description: Holiday ID

    delete:
      tags: [AdminHolidays]
      summary: Remove a holiday (admin)
      description: |
        Admin-only. Deletes the holiday record. All associated salon-closure rows are
        removed automatically via ON DELETE CASCADE.
      operationId: removeHoliday
      responses:
        "204":
          description: Holiday removed

  # ─── Admin — Website ──────────────────────────────────────────────────────────
...
  /api/salon/{salonId}/holidays:
    parameters:
      - $ref: "#/components/parameters/SalonId"

    get:
      tags: [Booking]
      summary: List salon holidays (customer)
      description: >
        Returns all holidays defined for this salon. The public website uses this
        to label days as holidays in the opening-hours display.
      operationId: listHolidaysPublic
      responses:
        "200":
          description: List of holidays
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SalonHoliday"

  # ─── Customer — Closures (read-only) ─────────────────────────────────────────
  /api/salon/{salonId}/closures:
    parameters:
      - $ref: "#/components/parameters/SalonId"

    get:
      tags: [Booking]
      summary: List salon closures (customer)
      description: >
        Returns all closures defined for this salon. The booking wizard uses this
        to disable closed date ranges in all three calendar views.
      operationId: listClosuresPublic
      responses:
        "200":
          description: List of closures
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SalonClosure"

  # ─── Admin — Closures ────────────────────────────────────────────────────────
  /api/salon-admin/{salonId}/closures:
    parameters:
      - $ref: "#/components/parameters/SalonId"

    get:
      tags: [AdminClosures]
      summary: List salon closures (admin)
      description: Returns all closures defined for this salon, sorted by start date.
      operationId: listClosures
      responses:
        "200":
          description: List of closures
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SalonClosure"

    post:
      tags: [AdminClosures]
      summary: Add a salon closure (admin)
      description: |
        Admin-only. Marks the salon as closed for a date range. A single day can be closed
        by setting `startDate` and `endDate` to the same value. Once saved, slot discovery
        will return an empty list for any date within the closure, and booking creation will
        return HTTP 400 for those dates.
      operationId: addClosure
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddClosureRequest"
      responses:
        "200":
          description: Closure created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalonClosure"
  /api/salon-admin/{salonId}/closures/{closureId}:
    parameters:
      - $ref: "#/components/parameters/SalonId"
      - name: closureId
        in: path
        required: true
        schema:
          type: integer
          format: int64
        description: Closure ID

    delete:
      tags: [AdminClosures]
      summary: Remove a salon closure (admin)
      description: Admin-only. Removes the closure so bookings can be taken for those dates again.
      operationId: removeClosure
      responses:
        "204":
          description: Closure removed

  # ─── Super Admin ──────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT issued by the identity provider after the admin authenticates.
        **Not yet enforced** — authentication/authorisation is a planned future addition.
        Currently the `salonId` path parameter is the only enforcement in place.
  parameters:
    SalonId:
      name: salonId
      in: path
      required: true
      schema:
        type: string
      description: |
        Salon UUID (`a1b2c3d4-e5f6-7890-abcd-ef1234567890`) or handler slug (`glam-salon`).
        UUID form is tried first; if the value isn't a valid UUID it falls back to a handler
        lookup — the same rule `GET /api/salon/{salonIdOrHandler}` uses. Returns `404` if
        neither matches.
...
  responses:
    NotFound:
      description: Resource not found
  schemas:
...
    SalonClosure:
      type: object
      properties:
        id:
          type: integer
          format: int64
        salonId:
          type: string
          format: uuid
        startDate:
          type: string
          format: date
          example: "2026-08-01"
        endDate:
          type: string
          format: date
          example: "2026-08-14"
        reason:
          type: string
          example: "Annual vacation"
        holidayId:
          type: integer
          format: int64
          nullable: true
          description: >
            Set when this closure was auto-generated from a holiday record.
            Holiday-backed closures cannot be deleted directly — delete the parent holiday instead.
    SalonHoliday:
      type: object
      properties:
        id:
          type: integer
          format: int64
        salonId:
          type: string
          format: uuid
        name:
          type: string
          example: "Christmas Day"
        month:
          type: integer
          minimum: 1
          maximum: 12
          description: Month of the holiday (1 = January)
        day:
          type: integer
          minimum: 1
          maximum: 31
          description: Day-of-month of the holiday
        year:
          type: integer
          nullable: true
          description: >
            Specific year for a one-off holiday; null for an annually recurring holiday.
        createdAt:
          type: string
          format: date-time
    AddHolidayRequest:
      type: object
      required: [name, month, day]
      properties:
        name:
          type: string
          description: Human-readable holiday name
          example: "Christmas Day"
        month:
          type: integer
          minimum: 1
          maximum: 12
        day:
          type: integer
          minimum: 1
          maximum: 31
        year:
          type: integer
          nullable: true
          description: >
            Omit (or send null) for an annually recurring holiday.
            Provide a specific year for a one-off occurrence.
    AddClosureRequest:
      type: object
      required: [startDate, endDate]
      properties:
        startDate:
          type: string
          format: date
          description: First day of the closure (inclusive)
        endDate:
          type: string
          format: date
          description: Last day of the closure (inclusive); use same value as startDate for a single day
        reason:
          type: string
          description: Optional human-readable reason shown in the admin UI

    # ── Salon ────────────────────────────────────────────────────────────────
    Salon:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        handler:
          type: string
          description: URL-friendly slug
        owner:
          $ref: "#/components/schemas/Owner"
        location:
          $ref: "#/components/schemas/Location"
        contact:
          $ref: "#/components/schemas/ContactInfo"
        operatingHours:
          type: array
          items:
            $ref: "#/components/schemas/OperatingHours"
        features:
          type: array
          items:
            $ref: "#/components/schemas/SalonFeature"
          description: Serialized as a flat array of feature strings
          example: ["BOOKING", "ANALYTICS"]
        bookingAdvanceDays:
          type: integer
          description: How many days in advance customers can book. Defaults to 60.
          default: 60
        businessRegistrationId:
          type: string
          description: Company/business registration number (e.g. CVR, EIN, Company Number).
        showBusinessId:
          type: boolean
          description: Whether to display the registration number on the public website. Defaults to false.
          default: false
        termsAccepted:
          type: boolean
          description: Whether the salon owner has accepted the terms and conditions. Defaults to false.
          default: false
        termsAcceptedAt:
          type: string
          format: date-time
          description: Timestamp when the terms were accepted. Set automatically on creation when termsAccepted is true.
          readOnly: true
        bookingRequiresConfirmation:
          type: boolean
          description: >
            When true, new bookings are created with status PENDING and must be manually confirmed
            by an admin. When false (default), bookings are auto-confirmed immediately (status CONFIRMED).
          default: false
        businessIdLabel:
          type: string
          description: >
            Country-specific label for the registration number (e.g. "CVR Number" for Denmark,
            "EIN" for United States). Derived automatically from location.country on create/update;
            read-only in responses.
          readOnly: true
        createdAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [ACTIVE, DISABLED]
          description: Salon lifecycle status. Defaults to ACTIVE on creation.
    CreateSalonRequest:
      type: object
      required: [name, ownerName, ownerEmail, termsAccepted]
      properties:
        name:
          type: string
        ownerName:
          type: string
        ownerEmail:
          type: string
          format: email
        ownerPhone:
          type: string
        location:
          $ref: "#/components/schemas/Location"
        contact:
          $ref: "#/components/schemas/ContactInfo"
        operatingHours:
          type: array
          items:
            $ref: "#/components/schemas/OperatingHours"
        features:
          type: array
          items:
            $ref: "#/components/schemas/SalonFeature"
        businessRegistrationId:
          type: string
          description: Company/business registration number.
        showBusinessId:
          type: boolean
          description: Whether to show the registration number on the public website. Defaults to false.
          default: false
        termsAccepted:
          type: boolean
          description: Must be true — the owner explicitly accepted the terms and conditions before salon creation.
    CreateSalonResponse:
      type: object
      properties:
        salonId:
          type: string
          format: uuid
          description: The newly created salon's UUID
        salonHandler:
          type: string
          description: URL-friendly slug derived from the salon name
        emailId:
          type: string
          format: email
          description: The owner's email address to use when signing in to the admin panel
        message:
          type: string
          description: Human-readable confirmation that a welcome email has been sent with the login link and setup guide
    UpdateSalonRequest:
      type: object
      required: [name]
      description: >
        Salon admin update — owner data is NOT changeable here.
        To change the owner, use PUT /api/salon-super-admin/salons/{id}/owner (super-admin only).
      properties:
        name:
          type: string
        location:
          $ref: "#/components/schemas/Location"
        contact:
          $ref: "#/components/schemas/ContactInfo"
        operatingHours:
          type: array
          items:
            $ref: "#/components/schemas/OperatingHours"
        bookingAdvanceDays:
          type: integer
          description: How many days ahead customers can book. Preserved if omitted.
        businessRegistrationId:
          type: string
          description: Business registration number (e.g. CVR for Denmark, EIN for USA). Displayed on the public website when showBusinessId is true.
        showBusinessId:
          type: boolean
          default: false
          description: Whether to display the business registration number on the public website. When updated alongside location.country, businessIdLabel is re-derived automatically.
        bookingRequiresConfirmation:
          type: boolean
          description: >
            When true, new bookings require manual admin confirmation (created as PENDING).
            When false, bookings are auto-confirmed. Preserved if omitted.
    PatchBookingSettingsRequest:
      type: object
      description: Partial update for booking-related salon settings. Omitted fields are preserved.
      properties:
        bookingAdvanceDays:
          type: integer
          description: How many days ahead customers can book.
          example: 60
        bookingRequiresConfirmation:
          type: boolean
          description: >
            When true, new bookings are created as PENDING and must be manually confirmed.
            When false, bookings are auto-confirmed immediately.
...
    Owner:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
    Location:
      type: object
      properties:
        address:
          type: string
        city:
          type: string
        state:
          type: string
        country:
          type: string
        zipCode:
          type: string
    ContactInfo:
      type: object
      properties:
        phone:
          type: string
        email:
          type: string
          format: email
        website:
          type: string
          format: uri
    OperatingHours:
      type: object
      properties:
        day:
          type: string
          enum: [MONDAY, TUESDAY, WEDNESDAY, THURSDAY, FRIDAY, SATURDAY, SUNDAY]
        openTime:
          type: string
          description: Opening time in HH:mm format
          example: "09:00"
        closeTime:
          type: string
          description: Closing time in HH:mm format
          example: "18:00"
        closed:
          type: boolean
          default: false
    SalonFeature:
      type: string
      enum:
        - STATIC_WEBSITE
        - BOOKING
        - MEMBERSHIP
        - WEBSHOP
        - ANALYTICS
        - LOYALTY_PROGRAM

    # ── Service ───────────────────────────────────────────────────────────────
...